+++ CHANGELOG.md
@@ -38,6 +38,11 @@ MarkLogic Rewriter XML Integration:
 1. Cache the endpoints to improve the line marker performance when the data has been previously calculated.
 2. In the Endpoints tool window navigate to the module file, not the rewriter XML element that references it.
 
+MarkLogic Search Options XML Integration:
+
+1. Support resolving `parse/@at`, `start-facet/@at`, and `finish-facet/@at` attribute values in XQuery direct
+   constructors to the XQuery/MJS/SJS file they reference.
+
 ## 2021
 
 *  [1.9.1 - 2021-07-23](docs/_posts/2021-07-23-release-1.9.1.md)